clean up jsonpath code; (hacky) support for array values
This moves the jsonpath code into functions which can be
refactored. Refactoring will most likelyl be necessary because of a
hack to determine whether the value or an array should be returned.

Given the following:
```javascript
{
  "a": 1,
  "b": [2,3,4,5],
  "c": [6],
}
```
this will return:
`$.a = 1`
`$.b[0] = 2`
`$.b[-2:] = [4, 5]`
`$.c = 6`

`$.c` should return `[6]` but since it's a single element array the
hacky solution will return the first element because there is no way
to detect whether the intention was to return a single-element array
or just a value becasue 3rd party modules always return everything
in a tuple.