refactor input/output processing code
Instead of processing input and output in the (multiprocess) executor it
has been moved to the workflow execution object.

With the refactor of input/output processing, the Pass task state can
now alter those values. This will allow for more complex workflows by
allowing the Pass state to be a data cleanup and processing step.